Simplified core mod.
